requirements.txt
torch
transformers
numpy
scikit-learn
pdf2image
pytesseract
Pillow
